Create animate_blanking.mcr
This script blanks line data for time steps greater than the current time step so that the line appears to grow with time. This macro is to be applied on CONVERGE data where .out files are also loaded in 2 separate XY Line plot frames. These 3 frames were tiled vertically where the top frame is titled "3D Data", the middle frame is titled "Line Data with Blanking", and the bottom frame is titled "Line Data with Marker". All of these frames are linked with time. A .mp4 animation is saved in the same location as this script.